+++ Scripts/ThirdPartyServices/AnalyticEvents/CommonEvents/LevelEnd.cs
@@ -7,10 +7,10 @@
         public int    level;
         public string status;
         public int    worldId;
-        public long   timePlay;
+        public double timePlay;
         public long   timerStamp;
 
-        public LevelEnd(int level, string status, int worldId, long timePlay, long timeStamp)
+        public LevelEnd(int level, string status, int worldId, double timePlay, long timeStamp)
         {
             this.level      = level;
             this.status     = status;
