+++ Scripts/ThirdPartyServices/AnalyticEvents/Mirai/MiraiAnalyticEventFactory.cs
@@ -87,11 +87,11 @@ namespace TheOneStudio.UITemplate.UITemplate.ThirdPartyServices.AnalyticEvents.M
 
         public override IEvent LevelStart(int level, int gold) { return new LevelStart(level, gold, 0, DateTimeOffset.UtcNow.ToUnixTimeSeconds()); }
 
-        public override IEvent LevelLose(int level, int timeSpent, int loseCount) { return new LevelEnd(level, "lose", 0, timeSpent, DateTimeOffset.UtcNow.ToUnixTimeSeconds()); }
+        public override IEvent LevelLose(int level, double timeSpent, int loseCount) { return new LevelEnd(level, "lose", 0, timeSpent, DateTimeOffset.UtcNow.ToUnixTimeSeconds()); }
 
-        public override IEvent LevelWin(int level, int timeSpent, int winCount) { return new LevelEnd(level, "win", 0, timeSpent, DateTimeOffset.UtcNow.ToUnixTimeSeconds()); }
+        public override IEvent LevelWin(int level, double timeSpent, int winCount) { return new LevelEnd(level, "win", 0, timeSpent, DateTimeOffset.UtcNow.ToUnixTimeSeconds()); }
 
-        public override IEvent LevelSkipped(int level, int timeSpent) { return new LevelEnd(level, "skip", 0, timeSpent, DateTimeOffset.UtcNow.ToUnixTimeSeconds()); }
+        public override IEvent LevelSkipped(int level, double timeSpent) { return new LevelEnd(level, "skip", 0, timeSpent, DateTimeOffset.UtcNow.ToUnixTimeSeconds()); }
 
         public override IEvent RewardedVideoShowFail(string place, string msg) { return new AdsRewardFail(place, msg); }
     }
